+++ src/app/(auth)/signUp/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import Link from 'next/link';
-import AuthForm from '@/components/shared/AuthForm';
+import AuthForm from '@/components/auth/AuthForm';
 
 export default function page() {
   return (
