+++ src/components/chat/EditAlert.tsx
@@ -1,63 +1,14 @@
 'use client';
 
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from '@/components/ui/alert-dialog';
+import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
-import { Imessage, useMessage } from '@/lib/stores/useMessagesStore';
+import { Input } from '@/components/ui/input';
+import { useMessage, Imessage } from '@/lib/stores/useMessagesStore';
 import browserClient from '@/utils/supabase/client';
 import { toast } from 'sonner';
-import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
-import { Input } from '@/components/ui/input';
 import React, { useRef } from 'react';
 
-export function DeleteAlert() {
-  const actionMessage = useMessage((state) => state.actionMessage);
-  const optimisticDeleteMessage = useMessage((state) => state.optimisticDeleteMessage);
-
-  const handleDeleteMessage = async () => {
-    const supabase = browserClient;
-    optimisticDeleteMessage(actionMessage?.id!);
-    const { error } = await supabase.from('messages').delete().eq('id', actionMessage?.id!);
-
-    if (error) {
-      toast.error(error.message);
-    } else {
-      toast.success('삭제 완료');
-    }
-  };
-
-  return (
-    <AlertDialog>
-      <AlertDialogTrigger asChild>
-        <button
-          className='sr-only'
-          id='trigger-delete'
-        ></button>
-      </AlertDialogTrigger>
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>메세지를 삭제하시겠습니까?</AlertDialogTitle>
-          <AlertDialogDescription>메세지를 정말로 삭제하시겠습니까?</AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <AlertDialogCancel>Cancel</AlertDialogCancel>
-          <AlertDialogAction onClick={handleDeleteMessage}>Continue</AlertDialogAction>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
-  );
-}
-
-export function EditAlert() {
+export default function EditAlert() {
   const actionMessage = useMessage((state) => state.actionMessage);
   const optimisticUpdateMessage = useMessage((state) => state.optimisticUpdateMessage);
   const inputRef = useRef() as React.MutableRefObject<HTMLInputElement>;
