+++ src/components/chat/ListMessages.tsx
@@ -3,43 +3,33 @@
 import { Imessage, useMessage } from '@/lib/stores/useMessagesStore';
 import browserClient from '@/utils/supabase/client';
 import Message from './Message';
-import { DeleteAlert, EditAlert } from './MessageAction';
+import DeleteAlert from './DeleteAlert';
+import EditAlert from './EditAlert';
 import { useEffect, useRef, useState } from 'react';
 import { toast } from 'sonner';
 import { ArrowDown } from 'lucide-react';
 
 export default function ListMessages() {
-  // 스크롤 위치를 추적하는 Ref
-  const scrollRef = useRef() as React.MutableRefObject<HTMLDivElement>;
+  const scrollRef = useRef<HTMLDivElement>(null); // 스크롤 위치 추적
   const [userScrolled, setUserScrolled] = useState(false);
+  const { messages, addMessage, optimisticIds, optimisticDeleteMessage, optimisticUpdateMessage } = useMessage();
 
-  const { messages, addMessage, optimisticIds, optimisticDeleteMessage, optimisticUpdateMessage } = useMessage(
-    (state) => state,
-  );
-  const supabase = browserClient;
-
-  // Supabase 구독 관리
+  // Supabase 구독 설정
   useEffect(() => {
+    const supabase = browserClient;
+
     const channel = supabase
       .channel('chat-room')
       .on('postgres_changes', { event: 'INSERT', schema: 'public', table: 'messages' }, async (payload) => {
         if (!optimisticIds.includes(payload.new.id)) {
-          const { error, data } = await supabase.from('users').select('*').eq('id', payload.new.send_by).single();
+          const { data, error } = await supabase.from('users').select('*').eq('id', payload.new.send_by).single();
           if (error) {
             toast.error(error.message);
           } else {
-            const newMessage = {
-              ...payload.new,
-              users: data,
-            };
-            addMessage(newMessage as Imessage);
+            addMessage({ ...payload.new, users: data } as Imessage);
           }
         }
-
-        const scrollContainer = scrollRef.current;
-        if (scrollContainer.scrollTop < scrollContainer.scrollHeight - scrollContainer.clientHeight - 10) {
-          setUserScrolled(true);
-        }
+        checkUserScroll(scrollRef.current);
       })
       .on('postgres_changes', { event: 'DELETE', schema: 'public', table: 'messages' }, (payload) => {
         optimisticDeleteMessage(payload.old.id);
@@ -52,32 +42,33 @@ export default function ListMessages() {
     return () => {
       channel.unsubscribe();
     };
-  }, [messages]);
+  }, [optimisticIds, addMessage, optimisticDeleteMessage, optimisticUpdateMessage]);
 
-  // 메시지가 추가될 때 스크롤을 아래로 이동
+  // 스크롤 처리
   useEffect(() => {
-    const scrollContainer = scrollRef.current;
-    if (scrollContainer && !userScrolled) {
-      scrollContainer.scrollTop = scrollContainer.scrollHeight;
+    if (scrollRef.current && !userScrolled) {
+      scrollRef.current.scrollTop = scrollRef.current.scrollHeight;
     }
   }, [messages]);
 
-  // 스크롤 이벤트 처리
+  // 스크롤 이벤트 처리 함수
   const handleOnScroll = () => {
     const scrollContainer = scrollRef.current;
     if (scrollContainer) {
-      const isScroll = scrollContainer.scrollTop < scrollContainer.scrollHeight - scrollContainer.clientHeight - 10;
-      setUserScrolled(isScroll);
-
-      if (scrollContainer.scrollTop === scrollContainer.scrollHeight - scrollContainer.clientHeight) {
-        setUserScrolled(false);
-      }
+      const atBottom = scrollContainer.scrollTop === scrollContainer.scrollHeight - scrollContainer.clientHeight;
+      setUserScrolled(!atBottom);
     }
   };
 
-  // 스크롤을 하단으로 이동
+  const checkUserScroll = (scrollContainer: HTMLDivElement | null) => {
+    if (scrollContainer) {
+      const scrolled = scrollContainer.scrollTop < scrollContainer.scrollHeight - scrollContainer.clientHeight - 10;
+      setUserScrolled(scrolled);
+    }
+  };
+
   const scrollDown = () => {
-    scrollRef.current.scrollTop = scrollRef.current.scrollHeight;
+    scrollRef.current?.scrollTo({ top: scrollRef.current.scrollHeight, behavior: 'smooth' });
   };
 
   return (
@@ -95,7 +86,6 @@ export default function ListMessages() {
             />
           ))}
         </div>
-
         <DeleteAlert />
         <EditAlert />
       </div>
@@ -103,7 +93,7 @@ export default function ListMessages() {
       {userScrolled && (
         <div className='absolute bottom-20 w-full'>
           <div
-            className='w-10 h-10 bg-white rounded-full flex justify-center items-center border mx-auto cursor-pointer '
+            className='w-10 h-10 bg-white rounded-full flex justify-center items-center border mx-auto cursor-pointer'
             onClick={scrollDown}
           >
             <ArrowDown />
